Enable require.js mode.
If the yml file contains a `require_js` element with the url fragment to require.js, then the server will switch to CommonJS amd module mode.

It will no longer directly load any source files, but will expect that the specs are declared as modules taking dependencies on the appropriate source files.