Pinning rack versions in Gemfile for Travis
The search space seems to be too big with our non-constrained rack
versions. Adding in these constraints just for the Gemfile seems to get
the build passed the 'resolving dependency step'.

We didn't add these constraints to the gemspec because it seems to work
when other gems don't have a rack dependency. It also seems to work when
a Gemfile.lock exists (which should be for most of our users).