Fix Action View deprecation warnings in Rails 6
In a future version of Rails, ActionView::Base#initialize will no longer
have default values for the first three arguments of lookup_context,
assignments and controller. To fix this create the instance by using a
ActionView::LookupContext with an empty view path set, an empty
assignment hash and a controller using an anonymous controller class.